- fix: system tray crash and icon
