+++ apps/shinkai-desktop/src-tauri/tauri.conf.json
@@ -43,9 +43,6 @@
   },
   "app": {
     "withGlobalTauri": false,
-    "trayIcon": {
-      "iconPath": "icons/icon.png"
-    },
     "security": {
       "csp": null
     },
